extracted toCode method
