+++ backend/src/main/java/edu/kit/quak/core/circuit/model/layer/operation/ElementaryQuantumGate.java
@@ -1,8 +1,6 @@
 package edu.kit.quak.core.circuit.model.layer.operation;
 
 import edu.kit.quak.core.circuit.exceptions.InvalidOperationConfigurationException;
-import edu.kit.quak.core.circuit.model.QuantumCircuit;
-import edu.kit.quak.core.circuit.model.layer.operation.library.ConcreteQuantumOperation;
 import edu.kit.quak.core.circuit.model.layer.operation.library.QuantumOperationLibrary;
 import java.util.List;
 import lombok.Getter;
@@ -29,103 +27,6 @@ public class ElementaryQuantumGate extends QuantumOperation {
                 "Operation type mismatch: expected %s but got %s".formatted(getClass(), operationDefinition.getDefinition().getType())
             );
         }
-    }
-
-    @Override
-    protected String operatorToCode(QuantumCircuit quantumCircuit) {
-        String operatorCode = super.operatorToCode(quantumCircuit);
-        if (operationDefinition.getDefinition() instanceof ConcreteQuantumOperation<?> definition && definition.isHasRotationAngle()) {
-            return operatorCode + "(" + formatAngle(rotationAngle) + ")";
-        }
-        return operatorCode;
-    }
-
-    /** Named constants the QASM parser understands, with a small tolerance for round-trip matching. */
-    private static final double CONSTANT_MATCH_EPSILON = 1e-9;
-
-    /**
-     * Formatiert den Winkel für QASM. Die benannten Konstanten tau und euler sowie rationale Vielfache
-     * von pi werden symbolisch ausgegeben (z.B. "tau", "euler", "pi/2", "-pi/4", "2*pi/3"), ansonsten als
-     * Dezimalzahl. So überleben sie einen parse → toCode → parse Round-Trip.
-     */
-    private static String formatAngle(double angle) {
-        if (!Double.isFinite(angle)) {
-            // Guard against emitting non-QASM tokens like "Infinity"/"NaN"; fall back to a neutral angle.
-            return "0";
-        }
-        if (angle == 0.0) {
-            return "0";
-        }
-        // Check tau before the pi logic so that 2*pi is emitted as "tau" rather than "2*pi".
-        String namedConstant = tryFormatAsNamedConstant(angle);
-        if (namedConstant != null) {
-            return namedConstant;
-        }
-        String piTerm = tryFormatAsPiMultiple(angle);
-        if (piTerm != null) {
-            return piTerm;
-        }
-        if (angle == Math.rint(angle)) {
-            return Long.toString((long) angle);
-        }
-        return Double.toString(angle);
-    }
-
-    /** Emits "tau"/"euler" (and their negatives) for the matching constant values, otherwise {@code null}. */
-    private static String tryFormatAsNamedConstant(double angle) {
-        if (Math.abs(angle - Math.TAU) < CONSTANT_MATCH_EPSILON) {
-            return "tau";
-        }
-        if (Math.abs(angle + Math.TAU) < CONSTANT_MATCH_EPSILON) {
-            return "-tau";
-        }
-        if (Math.abs(angle - Math.E) < CONSTANT_MATCH_EPSILON) {
-            return "euler";
-        }
-        if (Math.abs(angle + Math.E) < CONSTANT_MATCH_EPSILON) {
-            return "-euler";
-        }
-        return null;
-    }
-
-    /**
-     * Versucht, den Winkel als rationales Vielfaches von pi darzustellen (Nenner bis 12).
-     * Liefert {@code null}, wenn keine passende Darstellung gefunden wird.
-     */
-    private static String tryFormatAsPiMultiple(double angle) {
-        final double epsilon = 1e-9;
-        double ratio = angle / Math.PI;
-        for (int denominator = 1; denominator <= 12; denominator++) {
-            double scaled = ratio * denominator;
-            long numerator = Math.round(scaled);
-            if (numerator != 0 && Math.abs(scaled - numerator) < epsilon) {
-                long gcd = gcd(Math.abs(numerator), denominator);
-                return buildPiTerm(numerator / gcd, denominator / gcd);
-            }
-        }
-        return null;
-    }
-
-    private static String buildPiTerm(long numerator, long denominator) {
-        StringBuilder sb = new StringBuilder();
-        if (numerator < 0) {
-            sb.append("-");
-            numerator = -numerator;
-        }
-        sb.append(numerator == 1 ? "pi" : numerator + "*pi");
-        if (denominator != 1) {
-            sb.append("/").append(denominator);
-        }
-        return sb.toString();
-    }
-
-    private static long gcd(long a, long b) {
-        while (b != 0) {
-            long temp = b;
-            b = a % b;
-            a = temp;
-        }
-        return a;
     }
 
     @Override
