+++ backend/src/main/java/edu/kit/quak/core/circuit/model/layer/operation/library/QuantumOperationLibrary.java
@@ -1,6 +1,5 @@
 package edu.kit.quak.core.circuit.model.layer.operation.library;
 
-import edu.kit.quak.core.circuit.model.QuantumCircuit;
 import edu.kit.quak.core.circuit.model.layer.operation.ElementaryQuantumGate;
 import edu.kit.quak.core.circuit.model.layer.operation.Measurement;
 import edu.kit.quak.core.circuit.model.layer.operation.QuantumOperation;
@@ -45,25 +44,5 @@ public enum QuantumOperationLibrary {
     @Override
     public String toString() {
         return name();
-    }
-
-    public String toCode(QuantumCircuit quantumCircuit) {
-        return switch (this) {
-            case H -> "h";
-            case X -> "x";
-            case Y -> "y";
-            case Z -> "z";
-            case CX -> "cx";
-            case CZ -> "cz";
-            case SWAP -> "swap";
-            case CCX -> "ccx";
-            case S -> "s";
-            case T -> "t";
-            // TODO Für Rotations-Gates musst du später vermutlich Parameter anhängen:
-            case RX -> "rx";
-            case RY -> "ry";
-            case RZ -> "rz";
-            case MEASURE -> "measure";
-        };
     }
 }
